fix: do not set programming selector viewType to null
Fixes #808
